+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Clarify that the admin page only works for Mastodon and Pleroma instances.
+- Added a warning on mobile devices suggesting to view the site on a larger computer.
 
 ### Changed